+++ core/src/main/java/cucumber/api/SummaryPrinter.java
@@ -1,0 +1,5 @@
+package cucumber.api;
+
+public interface SummaryPrinter {
+    public void print(cucumber.runtime.Runtime runtime);
+}