+++ core/src/main/java/cucumber/runtime/formatter/PluginFactory.java
@@ -1,8 +1,14 @@
 package cucumber.runtime.formatter;
 
+import cucumber.api.StepDefinitionReporter;
+import cucumber.api.SummaryPrinter;
 import cucumber.runtime.CucumberException;
+import cucumber.runtime.DefaultSummaryPrinter;
+import cucumber.runtime.NullSummaryPrinter;
 import cucumber.runtime.io.URLOutputStream;
 import cucumber.runtime.io.UTF8OutputStreamWriter;
+import gherkin.formatter.Formatter;
+import gherkin.formatter.Reporter;
 
 import java.io.File;
 import java.io.IOException;
@@ -53,6 +59,8 @@ public class PluginFactory {
         put("json", CucumberJSONFormatter.class);
         put("usage", UsageFormatter.class);
         put("rerun", RerunFormatter.class);
+        put("default_summary", DefaultSummaryPrinter.class);
+        put("null_summary", NullSummaryPrinter.class);
     }};
     private static final Pattern PLUGIN_WITH_FILE_PATTERN = Pattern.compile("([^:]+):(.*)");
     private String defaultOutFormatter = null;
@@ -151,7 +159,7 @@ public class PluginFactory {
         }
     }
 
-    private <T> Class<T> pluginClass(String pluginName) {
+    private static <T> Class<T> pluginClass(String pluginName) {
         Class<T> pluginClass = (Class<T>) PLUGIN_CLASSES.get(pluginName);
         if (pluginClass == null) {
             pluginClass = loadClass(pluginName);
@@ -160,7 +168,7 @@ public class PluginFactory {
     }
 
     @SuppressWarnings("unchecked")
-    private <T> Class<T> loadClass(String className) {
+    private static <T> Class<T> loadClass(String className) {
         try {
             return (Class<T>) Thread.currentThread().getContextClassLoader().loadClass(className);
         } catch (ClassNotFoundException e) {
@@ -181,5 +189,40 @@ public class PluginFactory {
         } finally {
             defaultOut = null;
         }
+    }
+
+    public static boolean isFormatterName(String name) {
+        Class pluginClass = getPluginClass(name);
+        if (Formatter.class.isAssignableFrom(pluginClass) || Reporter.class.isAssignableFrom(pluginClass)) {
+            return true;
+        }
+        return false;
+    }
+
+    public static boolean isStepDefinitionResporterName(String name) {
+        Class pluginClass = getPluginClass(name);
+        if (StepDefinitionReporter.class.isAssignableFrom(pluginClass)) {
+            return true;
+        }
+        return false;
+    }
+
+    public static boolean isSummaryPrinterName(String name) {
+        Class pluginClass = getPluginClass(name);
+        if (SummaryPrinter.class.isAssignableFrom(pluginClass)) {
+            return true;
+        }
+        return false;
+    }
+
+    private static Class getPluginClass(String name) {
+        Matcher pluginWithFile = PLUGIN_WITH_FILE_PATTERN.matcher(name);
+        String pluginName;
+        if (pluginWithFile.matches()) {
+            pluginName = pluginWithFile.group(1);
+        } else {
+            pluginName = name;
+        }
+        return pluginClass(pluginName);
     }
 }